Move decodeHTMLEntities util into headless
